LayoutTests/imported/w3c/web-platform-tests/editing/other/inline-editing-host-has-placeholder-with-before-pseudo-element-expected.txt



PASS The editing host should have inserted text
FAIL The editing host should have no text assert_equals: expected "" but got "<br>"
